+++ pytest_pyodide/hook.py
@@ -18,12 +18,25 @@ RUNTIMES_AND_HOST = RUNTIMES + ["host"]
 RUNTIMES_NO_HOST = [f"{runtime}-no-host" for runtime in RUNTIMES]
 
 
-def _filter_runtimes(runtime: list[str]) -> tuple[bool, set[str]]:
+def _filter_runtimes(runtime: str) -> tuple[bool, set[str]]:
+    """Preprocess the given runtime commandline parameter
+
+    >>> _filter_runtimes("chrome")
+    (True, {'chrome'})
+    >>> _filter_runtimes("chrome-no-host")
+    (False, {'chrome'})
+    >>> _filter_runtimes("chrome-no-host, host, firefox")
+    (True, ...)
+    """
+
     # Always run host test, unless 'no-host' is given.
     run_host = True
 
+    # "chrome, firefox, node" ==> ["chrome", "firefox", "node"]
+    runtimes = [rt.strip() for rt in runtime.split(",")]
+
     # remove duplicates
-    runtime_set = set(runtime)
+    runtime_set = set(runtimes)
 
     runtime_filtered = set()
     for rt in runtime_set:
@@ -33,10 +46,14 @@ def _filter_runtimes(runtime: list[str]) -> tuple[bool, set[str]]:
 
         runtime_filtered.add(rt)
 
-    # If '--rt chrome-no-host --rt host' is given, we run host tests.
+    # If '--rt "chrome-no-host, host"' is given, we run host tests.
     run_host = run_host or ("host" in runtime_filtered)
 
     runtime_filtered.discard("host")
+
+    for rt in runtime_filtered:
+        if rt not in RUNTIMES:
+            raise ValueError(f"Invalid runtime: {rt}")
 
     return run_host, runtime_filtered
 
@@ -89,10 +106,8 @@ def pytest_addoption(parser):
         "--rt",
         "--runtime",
         dest="runtime",
-        nargs="+",
-        default=["node"],
-        choices=RUNTIMES_AND_HOST + RUNTIMES_NO_HOST,
-        help="Select runtime (default: %(default)s)",
+        default="node",
+        help="Select runtimes to run tests (default: %(default)s)",
     )