+++ CHANGELOG.md
@@ -11,6 +11,9 @@
   override `selenium_standalone` fixture with `selenium_standalone_refresh`.
   [#65](https://github.com/pyodide/pytest-pyodide/pull/65)
 
+- Breaking: `--runtime` commandline flag now requires runtimes to be comma-separated.
+  [#76](https://github.com/pyodide/pytest-pyodide/pull/76)
+
 - Add support for a custom `SimpleHTTPRequestHandler` class in the pytest
   webserver code, by passing the `handler_cls` parameter in the
   `spawn_web_server` function.